tasks/rsa/requirements.txt
aiohttp==1.3.3
async-timeout==1.1.0
chardet==2.3.0
gunicorn==19.7.0
multidict==2.1.4
pycrypto==2.6.1
yarl==0.9.8
